+++ packages/ethers/src/client.ts
@@ -248,7 +248,19 @@ export class Web3Modal extends Web3ModalScaffold {
         info: Info
         provider: Provider
       }) => {
-        if (id === ConstantsUtil.EIP6963_CONNECTOR_ID && info && provider) {
+        if (id === ConstantsUtil.INJECTED_CONNECTOR_ID) {
+          const InjectedProvider = ethersConfig.injected
+          if (!InjectedProvider) {
+            throw new Error('connectionControllerClient:connectInjected - provider is undefined')
+          }
+          try {
+            EthersStoreUtil.setError(undefined)
+            await InjectedProvider.request({ method: 'eth_requestAccounts' })
+            this.setInjectedProvider(ethersConfig)
+          } catch (error) {
+            EthersStoreUtil.setError(error)
+          }
+        } else if (id === ConstantsUtil.EIP6963_CONNECTOR_ID && info && provider) {
           try {
             EthersStoreUtil.setError(undefined)
             await provider.request({ method: 'eth_requestAccounts' })
@@ -277,6 +289,12 @@ export class Web3Modal extends Web3ModalScaffold {
       checkInstalled(ids) {
         if (!ids) {
           return Boolean(window.ethereum)
+        }
+
+        if (ethersConfig.injected) {
+          if (!window?.ethereum) {
+            return false
+          }
         }
 
         return ids.some(id => Boolean(window.ethereum?.[String(id)]))
@@ -475,6 +493,10 @@ export class Web3Modal extends Web3ModalScaffold {
       }
     }
 
+    if (ethersConfig.injected) {
+      this.checkActiveInjectedProvider(ethersConfig)
+    }
+
     if (ethersConfig.auth || ethersConfig.email) {
       this.syncAuthConnector(w3mOptions.projectId, ethersConfig.auth, ethersConfig.email)
     }
@@ -550,7 +572,7 @@ export class Web3Modal extends Web3ModalScaffold {
     localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
     EthersStoreUtil.reset()
 
-    if (providerType === 'eip6963') {
+    if (providerType === 'injected' || providerType === 'eip6963') {
       provider?.emit('disconnect')
     } else {
       const walletConnectProvider = provider as unknown as EthereumProvider
@@ -639,6 +661,18 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
+  private checkActiveInjectedProvider(config: ProviderType) {
+    const InjectedProvider = config.injected
+    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
+
+    if (InjectedProvider) {
+      if (walletId === ConstantsUtil.INJECTED_CONNECTOR_ID) {
+        this.setInjectedProvider(config)
+        this.watchInjected(config)
+      }
+    }
+  }
+
   private checkActiveCoinbaseProvider(config: ProviderType) {
     const CoinbaseProvider = config.coinbase as unknown as ExternalProvider
     const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
@@ -681,6 +715,23 @@ export class Web3Modal extends Web3ModalScaffold {
       EthersStoreUtil.setIsConnected(true)
       this.setAddress(WalletConnectProvider.accounts?.[0])
       this.watchWalletConnect()
+    }
+  }
+
+  private async setInjectedProvider(config: ProviderType) {
+    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
+    const InjectedProvider = config.injected
+
+    if (InjectedProvider) {
+      const { address, chainId } = await EthersHelpersUtil.getUserInfo(InjectedProvider)
+      if (address && chainId) {
+        EthersStoreUtil.setChainId(chainId)
+        EthersStoreUtil.setProviderType('injected')
+        EthersStoreUtil.setProvider(config.injected)
+        EthersStoreUtil.setIsConnected(true)
+        this.setAddress(address)
+        this.watchCoinbase(config)
+      }
     }
   }
 
@@ -766,6 +817,45 @@ export class Web3Modal extends Web3ModalScaffold {
     const accountsChangedHandler = async (accounts: string[]) => {
       if (accounts.length > 0) {
         await this.setWalletConnectProvider()
+      }
+    }
+
+    if (provider) {
+      provider.on('disconnect', disconnectHandler)
+      provider.on('accountsChanged', accountsChangedHandler)
+      provider.on('chainChanged', chainChangedHandler)
+    }
+  }
+
+  private watchInjected(config: ProviderType) {
+    const provider = config.injected
+
+    function disconnectHandler() {
+      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+      EthersStoreUtil.reset()
+
+      provider?.removeListener('disconnect', disconnectHandler)
+      provider?.removeListener('accountsChanged', accountsChangedHandler)
+      provider?.removeListener('chainChanged', chainChangedHandler)
+    }
+
+    function accountsChangedHandler(accounts: string[]) {
+      const currentAccount = accounts?.[0]
+      if (currentAccount) {
+        EthersStoreUtil.setAddress(getOriginalAddress(currentAccount) as Address)
+      } else {
+        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        EthersStoreUtil.reset()
+      }
+    }
+
+    function chainChangedHandler(chainId: string) {
+      if (chainId) {
+        const chain =
+          typeof chainId === 'string'
+            ? EthersHelpersUtil.hexStringToNumber(chainId)
+            : Number(chainId)
+        EthersStoreUtil.setChainId(chain)
       }
     }
 
@@ -1141,6 +1231,29 @@ export class Web3Modal extends Web3ModalScaffold {
             }
           }
         }
+      } else if (providerType === ConstantsUtil.INJECTED_CONNECTOR_ID && chain) {
+        const InjectedProvider = provider
+        if (InjectedProvider) {
+          try {
+            await InjectedProvider.request({
+              method: 'wallet_switchEthereumChain',
+              params: [{ chainId: EthersHelpersUtil.numberToHexString(chain.chainId) }]
+            })
+            EthersStoreUtil.setChainId(chain.chainId)
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          } catch (switchError: any) {
+            if (
+              switchError.code === EthersConstantsUtil.ERROR_CODE_UNRECOGNIZED_CHAIN_ID ||
+              switchError.code === EthersConstantsUtil.ERROR_CODE_DEFAULT ||
+              switchError?.data?.originalError?.code ===
+                EthersConstantsUtil.ERROR_CODE_UNRECOGNIZED_CHAIN_ID
+            ) {
+              await EthersHelpersUtil.addEthereumChain(InjectedProvider, chain)
+            } else {
+              throw new Error('Chain is not supported')
+            }
+          }
+        }
       } else if (providerType === ConstantsUtil.EIP6963_CONNECTOR_ID && chain) {
         const EIP6963Provider = provider
 
@@ -1224,6 +1337,21 @@ export class Web3Modal extends Web3ModalScaffold {
       })
     }
 
+    if (config.injected) {
+      const injectedConnectorType =
+        PresetsUtil.ConnectorTypesMap[ConstantsUtil.INJECTED_CONNECTOR_ID]
+      if (injectedConnectorType) {
+        w3mConnectors.push({
+          id: ConstantsUtil.INJECTED_CONNECTOR_ID,
+          explorerId: PresetsUtil.ConnectorExplorerIds[ConstantsUtil.INJECTED_CONNECTOR_ID],
+          imageId: PresetsUtil.ConnectorImageIds[ConstantsUtil.INJECTED_CONNECTOR_ID],
+          imageUrl: this.options?.connectorImages?.[ConstantsUtil.INJECTED_CONNECTOR_ID],
+          name: PresetsUtil.ConnectorNamesMap[ConstantsUtil.INJECTED_CONNECTOR_ID],
+          type: injectedConnectorType
+        })
+      }
+    }
+
     if (config.coinbase) {
       w3mConnectors.push({
         id: ConstantsUtil.COINBASE_CONNECTOR_ID,
@@ -1252,7 +1380,8 @@ export class Web3Modal extends Web3ModalScaffold {
         name: 'Auth',
         provider: this.authProvider,
         email,
-        socials: auth?.socials
+        socials: auth?.socials,
+        showWallets: auth?.showWallets === undefined ? true : auth.showWallets
       })
 
       super.setLoading(true)