+++ packages/scaffold/src/partials/w3m-email-login-widget/index.ts
@@ -46,8 +46,8 @@ export class W3mEmailLoginWidget extends LitElement {
 
   // -- Render -------------------------------------------- //
   public override render() {
-    const multipleConnectors = this.connectors.length > 1
     const connector = this.connectors.find(c => c.type === 'AUTH')
+    const multipleConnectors = this.connectors.length > 1
 
     if (!connector?.email) {
       return null
@@ -67,7 +67,11 @@ export class W3mEmailLoginWidget extends LitElement {
         <input type="submit" hidden />
       </form>
 
-      ${multipleConnectors ? html`<wui-separator text="or"></wui-separator>` : null}
+      ${connector.socials || !multipleConnectors
+        ? null
+        : html`<wui-flex .padding=${['xxs', '0', '0', '0'] as const}>
+            <wui-separator text="or"></wui-separator>
+          </wui-flex>`}
     `
   }