+++ packages/ui/src/components/wui-shimmer/styles.ts
@@ -3,7 +3,7 @@ import { css } from 'lit'
 export default css`
   :host {
     display: block;
-    box-shadow: inset 0 0 0 1px var(--wui-gray-glass-005);
+    box-shadow: inset 0 0 0 1px var(--wui-color-gray-glass-005);
     background: linear-gradient(
       120deg,
       var(--wui-color-bg-200) 5%,