+++ packages/ui/src/composites/wui-list-button/styles.ts
@@ -5,7 +5,7 @@ export default css`
     column-gap: var(--wui-spacing-s);
     padding: 16.5px var(--wui-spacing-l) 16.5px var(--wui-spacing-xs);
     width: 100%;
-    background-color: var(--wui-gray-glass-002);
+    background-color: var(--wui-color-gray-glass-002);
     border-radius: var(--wui-border-radius-xs);
     color: var(--wui-color-fg-100);
     justify-content: center;
@@ -13,7 +13,7 @@ export default css`
   }
 
   button:disabled {
-    background-color: var(--wui-gray-glass-015);
-    color: var(--wui-gray-glass-015);
+    background-color: var(--wui-color-gray-glass-015);
+    color: var(--wui-color-gray-glass-015);
   }
 `