+++ packages/ui/src/composites/wui-profile-button/index.ts
@@ -21,11 +21,9 @@ export class WuiProfileButton extends LitElement {
 
   @property() public avatarSrc?: string = undefined
 
-  @property({ type: Boolean }) public isProfileName = false
+  @property() public profileName?: string = ''
 
   @property() public address = ''
-
-  @property() public profileName?: string = undefined
 
   @property() public icon: IconType = 'chevronBottom'
 
@@ -42,7 +40,7 @@ export class WuiProfileButton extends LitElement {
         <wui-flex gap="xs" alignItems="center">
           <wui-text variant="large-600" color="fg-100">
             ${UiHelperUtil.getTruncateString({
-              string: this.profileName ?? this.address,
+              string: this.profileName || this.address,
               charsStart: this.profileName ? 18 : 4,
               charsEnd: this.profileName ? 0 : 4,
               truncate: this.profileName ? 'end' : 'middle'