+++ packages/ui/src/composites/wui-snackbar/styles.ts
@@ -7,7 +7,7 @@ export default css`
     align-items: center;
     padding: var(--wui-spacing-xs) var(--wui-spacing-m) var(--wui-spacing-xs) var(--wui-spacing-xs);
     border-radius: var(--wui-border-radius-3xl);
-    border: 1px solid var(--wui-gray-glass-005);
+    border: 1px solid var(--wui-color-gray-glass-005);
     box-sizing: border-box;
     max-height: 40px;
     background-color: var(--wui-color-bg-175);