+++ packages/scaffold-vue/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @web3modal/scaffold-vue
 
+## 4.2.1
+
+### Patch Changes
+
+- Hotfix to support injected and announced wallets in in app browsers
+
+- Updated dependencies []:
+  - @web3modal/scaffold@4.2.1
+
 ## 4.2.0
 
 ### Patch Changes